move table to dedicated component
