+++ app/(saas)/settings/billing-profiles/[id]/coworkers/_features/coworkers-table/coworkers-table.types.ts
@@ -1,0 +1,3 @@
+export type CoworkersTableProps = {
+  id: string;
+};
